mobilenetv4_medium/requirements.txt
# ================================
# Core Deep Learning Framework
# ================================

torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0


# ================================
# CLIP / SigLIP / EVA / MobileCLIP
# ================================

open_clip_torch>=2.24.0
transformers>=4.42.0
timm>=1.0.7

sentencepiece>=0.2.0
ftfy>=6.1.1


# ================================
# Numerical / Scientific
# ================================

numpy>=1.26.0
scipy>=1.12.0


# ================================
# Image Processing
# ================================

pillow>=10.0.0
opencv-python>=4.10.0


# ================================
# Dataset / Training Utilities
# ================================

tqdm>=4.66.0
pyyaml>=6.0.0
pandas>=2.2.0


# ================================
# ONNX Export
# ================================

onnx>=1.16.0
onnxruntime>=1.18.0


# ================================
# Qualcomm AI Hub
# ================================

qai-hub>=0.12.0


# ================================
# HuggingFace Hub (model download)
# ================================

huggingface_hub>=0.23.0


# ================================
# Optional but strongly recommended
# ================================

matplotlib>=3.9.0
scikit-learn>=1.5.0


# ================================
# Performance / Stability
# ================================

packaging>=24.0
requests>=2.32.0